# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Servimach API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  client: 'client',
  professional: 'professional',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  replId: string;
  name: string;
  /** @nullable */
  email: string | null;
  role: UserRole;
  /** @nullable */
  profilePhoto?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  /** @nullable */
  location?: string | null;
  createdAt: string;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  client: 'client',
  professional: 'professional',
} as const;

export interface UserUpdate {
  name?: string;
  role?: UserUpdateRole;
  whatsapp?: string;
  location?: string;
}

export interface Professional {
  id: number;
  userId: number;
  replId?: string;
  name: string;
  profession: string;
  category: string;
  /** @nullable */
  description?: string | null;
  location: string;
  /** @nullable */
  profilePhoto?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  rating: number;
  reviewCount: number;
  createdAt: string;
}

export interface ProfessionalUpdate {
  profession?: string;
  category?: string;
  description?: string;
  location?: string;
  whatsapp?: string;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  accepted: 'accepted',
  completed: 'completed',
  rejected: 'rejected',
} as const;

export type OrderPaymentStatus = typeof OrderPaymentStatus[keyof typeof OrderPaymentStatus];


export const OrderPaymentStatus = {
  pending: 'pending',
  paid: 'paid',
} as const;

export interface Order {
  id: number;
  clientId: number;
  professionalId: number;
  description: string;
  price: number;
  commission: number;
  professionalEarnings: number;
  status: OrderStatus;
  paymentStatus: OrderPaymentStatus;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  transactionId?: string | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  professionalName?: string | null;
  /** @nullable */
  professionalProfession?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface ProfessionalDashboard {
  totalOrders: number;
  pendingOrders: number;
  completedOrders: number;
  totalEarnings: number;
  platformCommission: number;
  rating: number;
  recentOrders: Order[];
}

export interface ClientDashboard {
  totalOrders: number;
  activeOrders: number;
  completedOrders: number;
  totalSpent: number;
  recentOrders: Order[];
}

export interface OrderInput {
  professionalId: number;
  /** @minLength 10 */
  description: string;
  /** @minimum 1 */
  price: number;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  accepted: 'accepted',
  rejected: 'rejected',
  completed: 'completed',
} as const;

export type OrderStatusUpdatePaymentStatus = typeof OrderStatusUpdatePaymentStatus[keyof typeof OrderStatusUpdatePaymentStatus];


export const OrderStatusUpdatePaymentStatus = {
  pending: 'pending',
  paid: 'paid',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
  paymentStatus?: OrderStatusUpdatePaymentStatus;
  paymentMethod?: string;
  transactionId?: string;
}

export interface Review {
  id: number;
  orderId: number;
  clientId: number;
  professionalId: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @nullable */
  comment?: string | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  clientPhoto?: string | null;
  createdAt: string;
}

export interface ReviewInput {
  orderId: number;
  professionalId: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  comment?: string;
}

export type NotificationType = typeof NotificationType[keyof typeof NotificationType];


export const NotificationType = {
  order_received: 'order_received',
  order_accepted: 'order_accepted',
  order_rejected: 'order_rejected',
  order_completed: 'order_completed',
  review_received: 'review_received',
} as const;

export interface Notification {
  id: number;
  userId: number;
  message: string;
  type: NotificationType;
  read: boolean;
  createdAt: string;
}

export interface AdminStats {
  totalUsers: number;
  totalProfessionals: number;
  totalClients: number;
  totalOrders: number;
  completedOrders: number;
  pendingOrders: number;
  totalRevenue: number;
  platformEarnings: number;
}

/**
 * Opaque session token — `Bearer <sid>`
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type ListProfessionalsParams = {
category?: string;
search?: string;
sortBy?: ListProfessionalsSortBy;
};

export type ListProfessionalsSortBy = typeof ListProfessionalsSortBy[keyof typeof ListProfessionalsSortBy];


export const ListProfessionalsSortBy = {
  rating: 'rating',
  newest: 'newest',
} as const;

export type ListOrdersParams = {
status?: ListOrdersStatus;
};

export type ListOrdersStatus = typeof ListOrdersStatus[keyof typeof ListOrdersStatus];


export const ListOrdersStatus = {
  pending: 'pending',
  accepted: 'accepted',
  completed: 'completed',
  rejected: 'rejected',
} as const;