# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Servimach API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for the current user (filtered by role)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, accepted, completed, rejected]
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new service order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (accept/reject/complete)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: Get notifications for current user
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
        "401":
          description: Unauthorized
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "401":
          description: Unauthorized
...
components:
...
  schemas:
...
    Order:
      type: object
      required: [id, clientId, professionalId, description, price, commission, professionalEarnings, status, paymentStatus, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        professionalId:
          type: integer
        description:
          type: string
        price:
          type: number
          format: float
        commission:
          type: number
          format: float
        professionalEarnings:
          type: number
          format: float
        status:
          type: string
          enum: [pending, accepted, completed, rejected]
        paymentStatus:
          type: string
          enum: [pending, paid]
        paymentMethod:
          type: ["string", "null"]
        transactionId:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
        professionalName:
          type: ["string", "null"]
        professionalProfession:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderInput:
      type: object
      required: [professionalId, description, price]
      properties:
        professionalId:
          type: integer
        description:
          type: string
          minLength: 10
        price:
          type: number
          format: float
          minimum: 1
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, rejected, completed]
        paymentStatus:
          type: string
          enum: [pending, paid]
        paymentMethod:
          type: string
        transactionId:
          type: string
...
    Notification:
      type: object
      required: [id, userId, message, type, read, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        message:
          type: string
        type:
          type: string
          enum: [order_received, order_accepted, order_rejected, order_completed, review_received]
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
...