# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Servimach API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderPaymentStatus } from './orderPaymentStatus';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  clientId: number;
  professionalId: number;
  description: string;
  price: number;
  commission: number;
  professionalEarnings: number;
  status: OrderStatus;
  paymentStatus: OrderPaymentStatus;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  transactionId?: string | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  professionalName?: string | null;
  /** @nullable */
  professionalProfession?: string | null;
  createdAt: Date;
  updatedAt?: Date;
}